Use current fiscal year for PO creation (API)
When creating a PO, if no fiscal_year is specified by the caller, use
the current fiscal year of the ordering agency.  This change primarily
affects the Open-ILS/src/support-scripts/acq_order_reader.pl script (and
any other direct API callers) which does not offer the user a chance to
set the fiscal year manually.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>