Hide Print Page button on Simplified Pull List
On the Simplified Pull List interface, there is a button for "Print Page"
that shows up in the upper right corner. The suggestion is to hide that
button to avoid library staff printing just what's visible instead of using
the proper "Print Pull List" button that's actually on the page.

Signed-off-by: Ben Shum <bshum@biblio.org>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>