TPAC: Display full publication info
In both the results table and record summary, we were failing to display
the place of publication - which is critical information for people
trying to track down the right edition or constructing a bibliography.

Grab the place of publication, and then reconstruct 260abc (avoiding
another XPath iteration and giving ourselves the ability to enhance the
schema.org microdata at the same time).

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Ben Shum <bshum@biblio.org>