LP#1047475 Set action.hold_request.email_notify to default to false
LP#1047475 points out that email notify prefs are ignored in TPAC.
Even if a patron deselects email notify during hold request time, it
will always be set to true due to the column default.

This change defaults the column in the database to false (like most
other similar columns are) and resolves the bug.

Credit to Steve Callender for the idea.

Signed-off-by: Michael Peters <mrpeters@library.in.gov>
Signed-off-by: Ben Shum <bshum@biblio.org>