lp1053026, more granular toolbar editing perms
client-side checking only to keep the honest folk out

Adds ADMIN_TOOLBAR_FOR_ORG, ADMIN_TOOLBAR_FOR_USER, and
ADMIN_TOOLBAR_FOR_WORKSTATION to the permission list

Staff needs at least one of these to create a toolbar, and they still need the
original ADMIN_TOOLBAR permission which is used on the server side of things.

For editing an existing toolbar, staff needs the perm corresponding to the
toolbar's ownership type (for example, if the toolbar is associated with a
user, then ADMIN_TOOLBAR_FOR_USER is needed).

For changing the ownership type of a toolbar one is allowed to edit, you
need the corresponding perm for the desired ownership type.

Signed-off-by: Jason Etheridge <jason@esilibrary.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>