LP1053526 - Don't inadvertently modify the in-memory org tree when printing
To see an example of this causing a problem, load Holdings Maintenance for
an arbitrary bib record with items. Make sure the Hide Empty Libs checkbox
is checked. Hit the Print button at the bottom of the UI (it's okay to then
Cancel the print). Then hit Refresh. It should result in an org.children()
is null error.

Signed-off-by: Jason Etheridge <jason@esilibrary.com>
Signed-off-by: Ben Shum <bshum@biblio.org>