Copy location group searching repairs
In the get_library BLOCK, upgrade any use of 'loc' to 'locg' instead of
trying to determine which to use in which situation.  Before this change,
'loc' was being used as the loc_name for the main org unit selector,
which should always use 'locg'.  For any other cases (related to search
params), it's always safe to use 'locg' over 'loc'.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>