Avoid specific version mention for database extensions
As tsbere pointed out, the 9.1 database script will be used for every
version after 9.1... which makes it confusing to have a version number.
As the primary reason for the difference is the switch to extensions
instead of contrib modules, use a name that flags that difference.

Convert the 9.1 full-text search configuration into a symbolic link
instead of maintaining a separate copy of the file.

Given that the pre-9.1 database script was focused on the creation of
PostgreSQL contribs, call it that, rather than the (now confusingly)
generic "create_database.sql" that existed from a time when contribs
ruled the earth and extensions were mere fantasy.

Adjust eg_db_config.sql for these changes as well - including the
command line switches.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Thomas Berezansky <tsbere@mvlc.org>