TPAC added content lookup uses local server address
Use the current Apache server's IP address instead of apache->hostname
to determine where to route TPAC added content lookups.  This allows the
server to handle the added content lookup locally instead of having to
route out and back into the cluster.

Continue using apache->hostname as the Host header to ensure the proper
virtualhost is used once the request is received.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Michael Peters <mrpeters@library.in.gov>