AutoFieldWidget support external copy loc retrieval
When rendering a collection of copy locations, AFW fetches locations
relevant to the context org unit.  While rendering an AFW for an
existing object, though, whose copy location is outside of that scope,
the code will now append the non-local copy location to the selector and
tag it with the owning lib (to avoid dupes).

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>