LP#1020261 Better label for "Host" field in acq.edi_account
The host field of acq.edi_account is really meant to contain more than just a hostname.
It needs a scheme part, so that the whole contents of the field should be URI-ish, like

ftp://ftp1.somevendor.com

Otherwise Evergreen's EDI mechanism can guess wrongly that we want to use
SSH instead of FTP to connect.

This also adds a description for the field, in the footer, along with the other
"helpers".

Signed-off-by: Michael Peters <mrpeters@library.in.gov>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>