LP1052941: Internal Server Error when adding to My List in staff client.
Don't redirect in EGCatLoader/Account.pm when in staff client context and
we're adding to a My List/Bookbag.

Add a wrapper macro to bookbag_actions.tt2 to wrap the add to book bag URLs
with some javascript that opens a new tab in the staff client. This is to
prevent the staff person from losing their search results as they would if
the new URL were opened in their current tab.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>
Signed-off-by: Ben Shum <bshum@biblio.org>