Add "Show Holds on Bib" menu items
Add a "Show holds on bibs" option from item checkin, item status, and patron
hold interfaces. This makes it quicker to flip to viewing who else has holds
on an item instead of having to click an option to see the bib, then clicking
in the menus for view holds. One less click.

Signed-off-by: Thomas Berezansky <tsbere@mvlc.org>
Signed-off-by: Ben Shum <bshum@biblio.org>