TPAC: Respect location visibility for result copies
Robert Soulliere noted in LP# 1062502 that copies in a location marked
"opac_visible = FALSE" were erroneously being displayed in search
results. This would occur if at least one copy on a record was in a
visible location, but other copies were in an invisible location.

This commit prevents copies in hidden locations from being displayed,
and also adds guards for the visibility of the copy status and the
circulation library.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Robert Soulliere <robert.soulliere@mohawkcollege.ca>