TPAC: First part of locg param can be branch shortname
Addresses LP #1020625.  You can use the locg param, which can set an
an OU for search context (and which has other implications related to
location/scope) by shortname now.

As long as your shortnames are something that work in an HTTP query
string, don't contain colons, and as long as you don't have numeric
shortnames for any orgs that would mask orgs with the sames numbers
as an ID (but that would be a lousy setup anyway), this should work.

Now case insensitive per recommendation by Michael Peters.  In the
event that you have org units with names that are the same except
for case (probably a bad practice), the locg parameter will simply
fail to set your context, rather than try to guess which org unit
you meant.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Pasi Kallinen <pasi.kallinen@pttk.fi>
Signed-off-by: Dan Scott <dscott@laurentian.ca>

Conflicts:
	Open-ILS/src/perlmods/lib/OpenILS/WWW/EGCatLoader/Util.pm