Warn on card expiring within 28 days
Thank you to Thomas Berzansky for pointing out a simple way of
accomplishing this, since we were already checking and alerting
if card was already expired.

This places an alert on the "stop sign" page if the card expires
within 28 days from the current date.

Interval can be configured by editing preexpire.setDate in display.js.

Signed-off-by: Michael Peters <mrpeters@library.in.gov>
Signed-off-by: Ben Shum <bshum@biblio.org>
Signed-off-by: Dan Scott <dscott@laurentian.ca>