Concerto test data: realistic call numbers
Add a parameter to the populate call number function so that we can define the
call number class. As a result, create a new 3-parm version of the function
that simply calls the 4-parm version with a NULL value for the 3rd parm.

Use some realistic LC and Dewey call numbers for the "concerto" bibs so
that we can have more useful tests of the call number browse functionality.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Bill Erickson <berick@esilibrary.com>