Add calculate_expire_time helper function to Holds.pm.
This new function calculates an expire_time for a hold based on the hold
expiration interval setting for a passed in org_unit. If the setting is found
the interval is added to "now" and returned as an ISO8601 string. undef is
returned if the setting is not found for the org_unit or its ancestors.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>
Signed-off-by: Ben Shum <bshum@biblio.org>