Resolve LP893448: Prevent hold expiration in past.
If a hold is placed with an expiration date in the past, the
expire_time of the hold will be reset using the hold expiration
interval settings.

Likewise, if a hold is changed to have an expiration date in the
past, the hold expiration date will be recalculated using the
hold expiration interval settings.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>
Signed-off-by: Ben Shum <bshum@biblio.org>