Tweak wording of patron account expiry warning
A warning is something that comes in advance of an event, so a
pre-warning would be a warning about the warning, which is a bit weird.
Hopefully this language will be a little simpler and easier for users to
understand.

Also, fetch the value of the setting once and stuff it in a variable;
vars are cheap and things might be a little more legible.

Also, add a release note entry for the new feature.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Ben Shum <bshum@biblio.org>