KPAC - fix getit_results with short titles
When using the "Get it" options to either add an item to a list or place an
item on hold, one would get an internal server error whenever attemping this
on a bib record with a short title.

Turns out that the substr function that shortens the length of long titles
breaks when the title is 18 characters or less.

Add an IF check around the substr function to avoid this error.

Signed-off-by: Ben Shum <bshum@biblio.org>
Signed-off-by: Michael Peters <mrpeters@library.in.gov>