LP#1082260: Update OPAC low search hits template for i18n
The OPAC lowhits.tt2 template is only partially translatable,
and the translatable part is impossible to make sound good
for languages with different word order than English.

This patch changes the already translatable part so there
are 4 different phrases to translate (instead of two phrase
fragments), and makes the search hints translatable.

For English language, the patch does not introduce any visible differences.

Signed-off-by: Pasi Kallinen <pasi.kallinen@pttk.fi>
Signed-off-by: Dan Scott <dscott@laurentian.ca>