LP#1078596: Cannot translate strings handled by fieldmapper
Strings that should be translatable only show up in English,
for example in the Patron Registration screen.

There's a bug in the fieldmapper where it sets a variable to
an empty array, but later in the code only checks whether that
particular variable exists, not if it's an empty array.

Signed-off-by: Pasi Kallinen <pasi.kallinen@pttk.fi>
Signed-off-by: Dan Scott <dscott@laurentian.ca>