Improve styling for 'delayed' lineitems in PO UI
When a lineitem is cancelled, but "keep_debits" is true, the lineitem is
in effect delayed instead of truly cancelled.  Make it more obvious in
the UI the difference between delayed and truly cancelled lineitems by
applying a row styling for delayed items (similar to other status row
styling) and always show the cancel reason (label) in the lineitem
display in the PO, next to the non-title attributes.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>