Serials: printing routing slips fails with Javascript error
The symptom (an error in the Javascript console about dijit.form.Button
and an empty page for routing slips that you try to print at batch
recieve) may or may not be affected by whether you're in a production
environment with the pre-built dojo 1.3.3 package installed or a
development environment without it.

This simple fix makes the problem go away in any case.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Ben Shum <bshum@biblio.org>